feat: スライド4完成

+++ rustlings/exercises/02_functions/functions2.rs
@@ -1,4 +1,5 @@
-fn call_me(num: usize) {
+// TODO: Add the missing type of the argument `num` after the colon `:`.
+fn call_me(num: i32) {
     for i in 0..num {
         println!("Ring! Call number {}", i + 1);
     }
